ScoreData and ScoreManager functioning
